% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.binomial}
\alias{average.power.binomial}
\title{Computer average power of many binomial comparisons}
\usage{
average.power.binomial(n, alpha, p1, p2, alternative)
}
\arguments{
\item{n}{per-group sample size}

\item{alpha}{p-value threshold}

\item{p1}{probability in one group}

\item{p2}{probability in other group}

\item{alternative}{one- or two-sided test}
}
\description{
Computer average power of many binomial comparisons
}
