% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.coxph}
\alias{average.power.coxph}
\title{Compute the average power of many Cox regression models for a given number of events, p-value threshold, vector of effect sizes (log hazard ratio),and variance of predictor variables}
\usage{
average.power.coxph(n, alpha, logHR, v)
}
\arguments{
\item{n}{per-group sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{logHR}{log hazard ratio (vector)}

\item{v}{variance of predictor variable (vector)}
}
\description{
Compute the average power of many Cox regression models for a given number of events, p-value threshold, vector of effect sizes (log hazard ratio),and variance of predictor variables
}
\examples{
average.power.coxph(n=30,alpha=0.05,logHR=log(rep(c(1,2),c(9900,100))),v=rep(1,10000))
}
