% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.signrank}
\alias{average.power.signrank}
\title{Compute average power of many signrank tests}
\usage{
average.power.signrank(n, alpha, p)
}
\arguments{
\item{n}{sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{p}{Pr(Y>X), as in Noether (JASA 1987)}
}
\description{
Compute average power of many signrank tests
}
