% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.signtest}
\alias{average.power.signtest}
\title{Compute average power of many sign tests}
\usage{
average.power.signtest(n, alpha, p)
}
\arguments{
\item{n}{sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{p}{Pr(Y>X), as in Noether (JASA 1987)}
}
\description{
Compute average power of many sign tests
}
