% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.t.test}
\alias{average.power.t.test}
\title{Compute average power of many t-tests}
\usage{
average.power.t.test(n, alpha, delta, sigma, type = "two.sample")
}
\arguments{
\item{n}{per-group sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{delta}{difference of population means (vector)}

\item{sigma}{standard deviation (vector or scalar)}

\item{type}{type of t-test: two.sample, one.sample}
}
\description{
Compute average power of many t-tests;Uses classical power formula for t-test;Assumes equal variance and sample size
}
