% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDRsamplesize2-package.R
\docType{package}
\name{FDRsamplesize2-package}
\alias{FDRsamplesize2}
\alias{FDRsamplesize2-package}
\title{FDRsamplesize2: Computing Power and Sample Size for the False Discovery Rate in Multiple Applications}
\description{
Defines a collection of functions to compute average power and sample size for studies that use the false discovery rate as the final measure of statistical significance.
}
\author{
\strong{Maintainer}: Yonghui Ni \email{Yonghui.Ni@STJUDE.ORG}

Authors:
\itemize{
  \item Stanley Pounds \email{Stanley.Pounds@STJUDE.ORG}
}

}
\keyword{internal}
