% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{find.sample.size}
\alias{find.sample.size}
\title{Determines the sample size needed to achieve a desired power while controlling the FDR at a specified level.}
\usage{
find.sample.size(alpha, pwr, avepow.func, n0 = 3, n1 = 6, max.its = 50, ...)
}
\arguments{
\item{alpha}{Fixed p-value threshold  (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{avepow.func}{an R function to compute average power}

\item{n0}{Lower limit for initial sample size range}

\item{n1}{Upper limit for initial sample size range}

\item{max.its}{Number of iterations}

\item{...}{Additional arguments to average power function}
}
\description{
Determines the sample size needed to achieve a desired power while controlling the FDR at a specified level.
}
