% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.binomial}
\alias{n.fdr.binomial}
\title{Sample size calculation for Binomial data}
\usage{
n.fdr.binomial(fdr, pwr, p1, p2, alternative = "two.sided", pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{p1}{probability in one group (vector)}

\item{p2}{probability in other group (vector)}

\item{alternative}{one- or two-sided test}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of two-group comparisons under Binomial distribution
}
