% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.negbin}
\alias{n.fdr.negbin}
\title{Sample size calculation for negative binomial data}
\usage{
n.fdr.negbin(fdr, pwr, log.fc, mu, sig, pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{log.fc}{log fold-change (vector), usual null hypothesis is log.fc=0}

\item{mu}{read depth per gene (vector, same length as log.fc)}

\item{sig}{coefficient of variation (CV) per gene (vector, same length as log.fc)}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of negative binomial comparisons
}
\references{
SN Hart, TM Therneau, Y Zhang, GA Poland, and J-P Kocher (2013). Calculating Sample Size Estimates for RNA Sequencing Data. Journal of Computational Biology 20: 970-978.
}
