% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.oneway}
\alias{n.fdr.oneway}
\title{Sample size calculation for one-way ANOVA}
\usage{
n.fdr.oneway(fdr, pwr, theta, k, pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{theta}{sum of ((group mean - overall mean)/stdev)^2 across all groups for each hypothesis test (vector)}

\item{k}{the number of groups to be compared}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of one-way ANOVA tests
}
