% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.signrank}
\alias{n.fdr.signrank}
\title{Sample size calculation for signed rank tests}
\usage{
n.fdr.signrank(fdr, pwr, p, pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{p}{Pr(Y>X), as in Noether (JASA 1987)}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of signed rank tests
}
\references{
Noether, Gottfried E (1987) Sample size determination for some common nonparametric tests. Journal of the American Statistical Association, 82:645-647.
}
