% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.signtest}
\alias{n.fdr.signtest}
\title{Sample size calculation for sign tests}
\usage{
n.fdr.signtest(fdr, pwr, p, pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{p}{Pr(Y>X), as in Noether (JASA 1987)}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of sign tests
}
\examples{
n.fdr.signtest(fdr=0.1,pwr=0.8,p=rep(c(0.8,0.5), c(100,9900)),pi0.hat="BH")
}
\references{
Noether, Gottfried E (1987) Sample size determination for some common nonparametric tests. Journal of the American Statistical Association, 82:645-647.
}
