% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{power.cox}
\alias{power.cox}
\title{Compute the power of a single-predictor Cox regression model}
\usage{
power.cox(n, alpha, logHR, v)
}
\arguments{
\item{n}{per-group sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{logHR}{log hazard ratio (vector)}

\item{v}{variance of predictor variable (vector)}
}
\value{
vector of power estimates for two-sided test
}
\description{
Use the formula of Hseih and Lavori (2000) to compute the power of a single-predictor Cox model
}
\references{
Hsieh, FY and Lavori, Philip W (2000) Sample-size calculations for the Cox proportional hazards regression model with nonbinary covariates. Controlled Clinical Trials 21(6):552-560.
}
