% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{power.fisher}
\alias{power.fisher}
\title{Compute Power for Fisher's exact test}
\usage{
power.fisher(p1, p2, n, alpha, alternative)
}
\arguments{
\item{p1}{probability in one group (vector)}

\item{p2}{probability in other group (vector)}

\item{n}{per-group sample size}

\item{alpha}{p-value threshold}

\item{alternative}{one- or two-sided test}
}
\description{
Compute Power for Fisher's exact test
}
