% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{power.signrank}
\alias{power.signrank}
\title{Compute power of signrank test}
\usage{
power.signrank(n, alpha, p)
}
\arguments{
\item{n}{sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{p}{Pr(Y>X), as in Noether (JASA 1987)}
}
\description{
Use the Noether (1987) formula to compute the power of the signrank test
}
\references{
Noether, Gottfried E (1987) Sample size determination for some common nonparametric tests. Journal of the American Statistical Association, 82:645-647.
}
