% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{average.power.fisher}
\alias{average.power.fisher}
\title{Compute average power of many Fisher's exact tests}
\usage{
average.power.fisher(p1, p2, n, alpha, alternative)
}
\arguments{
\item{p1}{probability of success in group1 (vector)}

\item{p2}{probability of success in group1 (vector)}

\item{n}{per-group sample size}

\item{alpha}{p-value threshold}

\item{alternative}{one- or two-sided test}
}
\description{
Compute average power of many Fisher's exact tests
}
