% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.coxph}
\alias{n.fdr.coxph}
\title{Sample size calculation for the Cox proportional hazards regression model}
\usage{
n.fdr.coxph(fdr, pwr, logHR, v, pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{logHR}{log hazard ratio (vector)}

\item{v}{variance of predictor variable (vector)}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find number of events needed to have a desired false discovery rate and power for a large number of Cox regression models with nonbinary covariates
}
\references{
Hsieh, FY and Lavori, Philip W (2000) Sample-size calculations for the Cox proportional hazards regression model with nonbinary covariates. Controlled Clinical Trials 21(6):552-560.
}
