% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.poisson}
\alias{n.fdr.poisson}
\title{Sample size calculation for Poisson data}
\usage{
n.fdr.poisson(fdr, pwr, rho, mu0, w = 1, type = "w", pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{rho}{Fold-change, usual null hypothesis is that rho=1 (vector)}

\item{mu0}{Average count in control group}

\item{w}{Ratio of total number of}

\item{type}{Type of test: "w" for Wald, "s" for score, "lw" for log-transformed Wald, "ls" for log-transformed score.}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of two-group comparisons under Poisson distribution
}
\references{
C-I Li, P-F Su, Y Guo, and Y Shyr (2013). Sample size calculation for differential expression analysis of RNA-seq data under Poisson distribution. Int J Comput Biol Drug Des 6(4). doi:10.1504/IJCBDD.2013.056830
}
