% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesize.R
\name{n.fdr.ttest}
\alias{n.fdr.ttest}
\title{Sample size calculation for t-tests}
\usage{
n.fdr.ttest(fdr, pwr, delta, sigma = 1, type = "two.sample", pi0.hat = "BH")
}
\arguments{
\item{fdr}{Desired FDR (scalar numeric)}

\item{pwr}{Power of single α-level test (scalar numeric)}

\item{delta}{difference of population means (vector)}

\item{sigma}{standard deviation (vector or scalar)}

\item{type}{type of t-test}

\item{pi0.hat}{Approximation method for null proportion}
}
\description{
Find the sample size needed to have a desired false discovery rate and power for a large number of t-tests
}
