% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{power.hart}
\alias{power.hart}
\title{Compute power for RNA-seq experiments assuming negative binomial distribution}
\usage{
power.hart(n, alpha, log.fc, mu, sig)
}
\arguments{
\item{n}{per-group sample size (scalar)}

\item{alpha}{p-value threshold (scalar)}

\item{log.fc}{log fold-change (vector), usual null hypothesis is log.fc=0}

\item{mu}{read depth per gene (vector, same length as log.fc)}

\item{sig}{coefficient of variation (CV) per gene (vector, same length as log.fc)}
}
\value{
vector of power estimates for the set of two-sided tests
}
\description{
Use the formula of Hart et al (2013) to compute power for comparing RNA-seq expression across two groups assuming a negative binomial distribution
}
\details{
This function is based on equation (1) of Hart et al (2013). It assumes a negative binomial model for RNA-seq read counts and equal sample size per group.
}
\references{
SN Hart, TM Therneau, Y Zhang, GA Poland, and J-P Kocher (2013). Calculating Sample Size Estimates for RNA Sequencing Data. Journal of Computational Biology 20: 970-978.
}
