% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_power.R
\name{power.li}
\alias{power.li}
\title{Compute Power for RNA-Seq Experiments Assuming Poisson Distribution}
\usage{
power.li(n, alpha, rho, mu0, w = 1, type = "w")
}
\arguments{
\item{n}{per-group sample size}

\item{alpha}{p-value threshold (scalar)}

\item{rho}{Fold-change, usual null hypothesis is that rho=1 (vector)}

\item{mu0}{Average count in control group}

\item{w}{Ratio of total number of}

\item{type}{Type of test: "w" for Wald, "s" for score, "lw" for log-transformed Wald, "ls" for log-transformed score}
}
\value{
vector of power estimates for two-sided tests
}
\description{
Use the formula of Li et al (2013) to compute power for comparing RNA-seq expression across two groups assuming the Poisson distribution
}
\details{
This function computes the power for each of a series of two-sided tests defined by the input parameters. The power is based on the sample size formulas in equations 10-13 of Li et al (2013). Also, note that the null.effect is set to 1 in the examples because the usual null hypothesis is that the fold-change = 1.
}
\references{
C-I Li, P-F Su, Y Guo, and Y Shyr (2013). Sample size calculation for differential expression analysis of RNA-seq data under Poisson distribution. Int J Comput Biol Drug Des 6(4). doi:10.1504/IJCBDD.2013.056830
}
